Split widget: metric tile
