+++ extension/devtools/app/lib/widgets/metric_tile.dart
@@ -1,0 +1,57 @@
+part of '../main.dart';
+
+class _MetricTile extends StatelessWidget {
+  const _MetricTile({
+    required this.label,
+    required this.value,
+    required this.trend,
+    required this.accent,
+    required this.icon,
+  });
+
+  final String label;
+  final String value;
+  final String trend;
+  final Color accent;
+  final IconData icon;
+
+  @override
+  Widget build(BuildContext context) {
+    return _GlassCard(
+      padding: const EdgeInsets.symmetric(horizontal: 18, vertical: 16),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Row(
+            children: [
+              Container(
+                width: 32,
+                height: 32,
+                decoration: BoxDecoration(
+                  color: accent.withValues(alpha: 0.2),
+                  borderRadius: BorderRadius.circular(12),
+                ),
+                child: Icon(icon, color: accent, size: 18),
+              ),
+              const Spacer(),
+              _GlassPill(
+                padding: const EdgeInsets.symmetric(
+                  horizontal: 10,
+                  vertical: 6,
+                ),
+                color: accent.withValues(alpha: 0.18),
+                child: Text(trend),
+              ),
+            ],
+          ),
+          const SizedBox(height: 12),
+          Text(label, style: Theme.of(context).textTheme.labelLarge),
+          const SizedBox(height: 6),
+          Text(value, style: Theme.of(context).textTheme.headlineSmall),
+          const SizedBox(height: 10),
+          _Sparkline(color: accent),
+        ],
+      ),
+    );
+  }
+}
